% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fWAR}
\alias{fWAR}
\title{fWAR values downloaded from FanGraphs.}
\format{A data frame with one row for each player-season and 11 columns:
\describe{
   \item{playerid}{the FanGraphs id of the player}
   \item{yearId}{the season}
   \item{Name}{the player's name}
   \item{fWAR}{the player's fWAR in that season}
   \item{fRAA_bat}{the Runs Above Average earned while batting}
   \item{fRAA_br}{the Runs Above Average earned while running the bases}
   \item{fRAA_field}{the Runs Above Average earned while fielding}
   \item{fWAR_pitch}{the Runs Above Average earned while pitching}
   \item{fRAR}{the Runs Above Replacement}
   \item{fRAA}{the Runs Above Average}
   \item{fRepl}{the replacement level, in runs}
}}
\source{
\url{http://fangraphs.com/}
}
\usage{
fWAR
}
\description{
A dataset containing fWAR values
}
\keyword{datasets}

