% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getrWAR.R
\name{getrWAR}
\alias{getrWAR}
\title{getrWAR}
\usage{
getrWAR()
}
\value{
a data.frame consisting of rWAR
}
\description{
Retrieves rWAR from Baseball-Reference.com
}
\details{
Retrieves daily rWAR figures from Baseball-Reference.com and stores it as a data.frame
}
\examples{
# Get data from yesterday
\dontrun{
rWAR = getrWAR()
}
data(rWAR)
hist(rWAR$rWAR)

# Leaders since 1954
library(dplyr)
modern = filter(rWAR, yearId >= 1954)
head(arrange(modern, desc(rWAR)), 20)

# Relationship between batting and fielding
plot(modern$rRAA_bat, modern$rRAA_field, cex=0.5)
}

