% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{idBP}
\alias{idBP}
\title{ID table from Baseball-Reference.com}
\format{A data frame with one row for each player and 9 columns:
\describe{
   \item{bats}{}
   \item{bbrefID}{}
   \item{retroID}{the person's Retrosheet id}
}}
\source{
\url{http://www.baseballprospectus.com/sortable/playerids/playerid_list.csv}
}
\usage{
idBP
}
\description{
A data frame contains player ID mappings
}
\seealso{
\code{\link[Lahman]{Master}}
}
\keyword{datasets}

