% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MLBAM2012}
\alias{MLBAM2012}
\alias{MLBAM2013}
\alias{MLBAM2014}
\alias{MLBAM2015}
\alias{MLBAM2016}
\alias{MLBAM2017}
\title{Play-by-play information downloaded from MLBAM GameDay.}
\format{A data frame with approximately 180,000 rows (differs from year to year) and 62 columns:
\describe{
   \item{pitcherId}{the MLBAM id of the pitcher}
   \item{batterId}{the MLBAM id of the batter}
}}
\source{
\url{http://gd2.mlb.com/}
}
\usage{
MLBAM2012

MLBAM2013

MLBAM2014

MLBAM2015
}
\description{
A dataset containing play-by-play data from MLBAM GameDay and processed by 
\code{openWAR}.
}
\keyword{datasets}

