% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openWARData-package.R
\docType{package}
\name{openWARData}
\alias{openWARData}
\alias{openWARData-package}
\title{openWARData}
\description{
This package contains the data associated with openWAR
}
\details{
A collection of data sets used with openWAR.
}
\author{
Ben Baumer and Gregory J. Matthews
}

