% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{idBP}
\alias{idBP}
\title{ID table from Baseball-Reference.com}
\format{A data frame with one row for each player and 31 columns:
\describe{
   \item{nameLast}{the person's last name}
   \item{nameFirst}{the person's first name}
   \item{bpId}{the person's Baseball Prospectus id}
   \item{davenportId}{the person's Davenport id}
   \item{mlbamId}{the person's MLBAM id}
   \item{retroId}{the person's Retrosheet id}
}}
\source{
\url{http://www.baseballprospectus.com/sortable/playerids/playerid_list.csv}
}
\usage{
idBP
}
\description{
A data frame contains player ID mappings
}
\seealso{
\code{\link[Lahman]{Master}}
}
\keyword{datasets}

